reorganize db_v2 utils
